Move tests from personal_dashboard_layout to private_sources_sidebar
