+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/components/layout/personal_dashboard_layout/personal_dashboard_layout.test.tsx
@@ -5,8 +5,6 @@
  * 2.0.
  */
 
-import '../../../../__mocks__/shallow_useeffect.mock';
-
 import { setMockValues } from '../../../../__mocks__/kea_logic';
 
 import React from 'react';
@@ -17,59 +15,27 @@ import { EuiCallOut } from '@elastic/eui';
 
 import { AccountHeader } from '..';
 
-import { SourceSubNav } from '../../../views/content_sources/components/source_sub_nav';
-
-import {
-  PRIVATE_CAN_CREATE_PAGE_TITLE,
-  PRIVATE_VIEW_ONLY_PAGE_TITLE,
-  PRIVATE_VIEW_ONLY_PAGE_DESCRIPTION,
-  PRIVATE_CAN_CREATE_PAGE_DESCRIPTION,
-} from '../../../views/content_sources/constants';
-import { ViewContentHeader } from '../../shared/view_content_header';
-
 import { PersonalDashboardLayout } from './personal_dashboard_layout';
 
 describe('PersonalDashboardLayout', () => {
-  const mockValues = {
-    account: { canCreatePersonalSources: true },
-  };
-
   const children = <p data-test-subj="TestChildren">test</p>;
-
-  beforeEach(() => {
-    setMockValues({ ...mockValues });
-  });
+  const sidebar = <p data-test-subj="TestSidebar">test</p>;
 
   it('renders', () => {
-    const wrapper = shallow(<PersonalDashboardLayout>{children}</PersonalDashboardLayout>);
+    const wrapper = shallow(
+      <PersonalDashboardLayout sidebar={sidebar}>{children}</PersonalDashboardLayout>
+    );
 
     expect(wrapper.find('[data-test-subj="TestChildren"]')).toHaveLength(1);
-    expect(wrapper.find(SourceSubNav)).toHaveLength(1);
+    expect(wrapper.find('[data-test-subj="TestSidebar"]')).toHaveLength(1);
     expect(wrapper.find(AccountHeader)).toHaveLength(1);
-  });
-
-  it('uses correct title and description when private sources are enabled', () => {
-    const wrapper = shallow(<PersonalDashboardLayout>{children}</PersonalDashboardLayout>);
-
-    expect(wrapper.find(ViewContentHeader).prop('title')).toEqual(PRIVATE_CAN_CREATE_PAGE_TITLE);
-    expect(wrapper.find(ViewContentHeader).prop('description')).toEqual(
-      PRIVATE_CAN_CREATE_PAGE_DESCRIPTION
-    );
-  });
-
-  it('uses correct title and description when private sources are disabled', () => {
-    setMockValues({ account: { canCreatePersonalSources: false } });
-    const wrapper = shallow(<PersonalDashboardLayout>{children}</PersonalDashboardLayout>);
-
-    expect(wrapper.find(ViewContentHeader).prop('title')).toEqual(PRIVATE_VIEW_ONLY_PAGE_TITLE);
-    expect(wrapper.find(ViewContentHeader).prop('description')).toEqual(
-      PRIVATE_VIEW_ONLY_PAGE_DESCRIPTION
-    );
   });
 
   it('renders callout when in read-only mode', () => {
     const wrapper = shallow(
-      <PersonalDashboardLayout readOnlyMode>{children}</PersonalDashboardLayout>
+      <PersonalDashboardLayout sidebar={sidebar} readOnlyMode>
+        {children}
+      </PersonalDashboardLayout>
     );
 
     expect(wrapper.find(EuiCallOut)).toHaveLength(1);
